+++ threedi_models_and_simulations/data_models/simulation_data_models.py
@@ -2,6 +2,7 @@
 # Copyright (C) 2023 by Lutra Consulting for 3Di Water Management
 from dataclasses import dataclass
 from datetime import datetime
+from typing import List
 
 from threedi_api_client.openapi import (CurrentStatus, FileBoundaryCondition,
                                         FileRasterLeakage,
@@ -94,6 +95,7 @@ class InitialConditions(SimulationElement):
     aggregation_method_groundwater: str = None
     saved_state: str = None
     initial_concentrations_2d: dict = None
+    initial_concentrations_1d: dict = None
 
 
 @dataclass
@@ -115,13 +117,20 @@ class DWF(SimulationElement):
 
 @dataclass
 class Breach(SimulationElement):
-    breach_id: str = None
+    breach_id: int = None
     width: float = None
-    duration_in_units: float = None
+    duration_till_max_depth: float = None
     offset: float = None
     discharge_coefficient_positive: float = None
     discharge_coefficient_negative: float = None
+    levee_material: str = None
     max_breach_depth: float = None
+
+
+@dataclass
+class Breaches(SimulationElement):
+    potential_breaches: List[Breach] = None
+    flowlines: List[Breach] = None
 
 
 @dataclass
@@ -204,7 +213,7 @@ class NewSimulation:
     laterals: Laterals = None
     substances: Substances = None
     dwf: DWF = None
-    breach: Breach = None
+    breaches: Breaches = None
     precipitation: Precipitation = None
     wind: Wind = None
     settings: Settings = None